requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
langchain==0.0.300
openai==0.27.0
psycopg2-binary==2.9.3
redis==4.0.2
requests==2.28.1
python-dotenv==0.19.0